Fixed hover behavior.
