+++ src/chartLibraries/dygraph/hover.js
@@ -1,8 +1,6 @@
 import { removeEvent, addEvent } from "dygraphs/src/dygraph-utils"
 
 export default chartUI => {
-  let listeners = []
-
   const mouseout = () => {
     chartUI.sdk.trigger("blur", chartUI.chart)
     chartUI.chart.trigger("blur")
@@ -13,14 +11,13 @@ export default chartUI => {
     chartUI.chart.trigger("hover")
   }
 
-  const highlight = (event, x, points, row, seriesName) => {
-    if (!seriesName) return
+  const listeners = [chartUI.on("mouseout", mouseout), chartUI.on("mouseover", mouseover)]
 
-    const { offsetX, offsetY } = event
-    const { column } = chartUI.getDygraph().getPropertiesForSeries(seriesName)
-    chartUI.sdk.trigger("hoverX", chartUI.chart, offsetX, offsetY, row, column)
-    chartUI.chart.trigger("hoverX", offsetX, offsetY, row, column)
-  }
+  addEvent(
+    chartUI.getDygraph().mouseEventElement_,
+    "mousemove",
+    chartUI.getDygraph().mouseMoveHandler_
+  )
 
   const destroy = () => {
     removeEvent(
@@ -29,26 +26,7 @@ export default chartUI => {
       chartUI.getDygraph().mouseMoveHandler_
     )
     listeners.forEach(listener => listener())
-    mouseout()
   }
 
-  const toggle = enabled => {
-    destroy()
-
-    if (!enabled) return
-
-    addEvent(
-      chartUI.getDygraph().mouseEventElement_,
-      "mousemove",
-      chartUI.getDygraph().mouseMoveHandler_
-    )
-
-    listeners = [
-      chartUI.on("highlightCallback", highlight),
-      chartUI.on("mouseout", mouseout),
-      chartUI.on("mouseover", mouseover),
-    ]
-  }
-
-  return { toggle, destroy }
+  return destroy
 }
