+++ src/components/dimensionValue.js
@@ -3,7 +3,7 @@ import { TextMicro } from "@netdata/netdata-ui/lib/components/typography"
 
 const DimensionValue = ({ chart, index, ...rest }) => {
   const getValue = () => {
-    const hover = chart.getAttribute("hover")
+    const hover = chart.getAttribute("hoverX")
     const { result } = chart.getPayload()
     const x = hover ? hover[0] : result.data.length - 1
     return chart.getConvertedValue(result.data[x][index + 1])
@@ -14,7 +14,7 @@ const DimensionValue = ({ chart, index, ...rest }) => {
   const unmount = useRef(false)
 
   useEffect(() => {
-    const remove = chart.onAttributeChange("hover", () => {
+    const remove = chart.onAttributeChange("hoverX", () => {
       setState(getValue())
     })
